fix: fixed algorithm, need to improve 500+ and updated makefile

+++ src/nodes/find.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   find.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vde-frei <vde-frei@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nivicius <nivicius@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 20:53:22 by vde-frei          #+#    #+#             */
-/*   Updated: 2023/12/02 22:11:57 by vde-frei         ###   ########.fr       */
+/*   Updated: 2023/12/07 01:33:31 by nivicius         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,13 +35,13 @@ t_stack	*find_highest(t_stack *stack)
 
 t_stack	*find_smallest(t_stack *stack)
 {
-	int		smallest;
+	long	smallest;
 	t_stack	*smallest_node;
 
 	smallest_node = NULL;
 	if (NULL == stack)
 		return (NULL);
-	smallest = INT_MAX;
+	smallest = LONG_MAX;
 	while (stack)
 	{
 		if (stack->value < smallest)
